+++ webextension/js/classes/chrome-native-settings.js
@@ -1,4 +1,4 @@
 export const chromeNativeSettingsStorageKey = '_nativeSettings';
 export async function getElectronSettings() {
-	return (await browser.storage.local.get([chromeNativeSettingsStorageKey])) ?? {};
+	return (await chrome.storage.local.get([chromeNativeSettingsStorageKey])) ?? {};
 }